Notification and Admin feature update

+++ app/src/main/java/com/example/qr/activities/EventListFragment.java
@@ -1,5 +1,6 @@
 package com.example.qr.activities;
 
+import android.app.AlertDialog;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -15,15 +16,25 @@ import androidx.fragment.app.Fragment;
 import com.example.qr.R;
 import com.example.qr.models.Event;
 import com.example.qr.models.EventArrayAdapter;
+import com.example.qr.models.Notification;
 import com.example.qr.utils.FirebaseUtil;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GetTokenResult;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 
 import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.messaging.FirebaseMessaging;
+
 
 /**
  * EventListFragment displays a list of events, allowing users to view detailed information or edit them.
@@ -34,7 +45,7 @@ public class EventListFragment extends Fragment {
     ArrayList<Event> eventDataList;
     EventArrayAdapter eventArrayAdapter;
     String message;
-    Map<String, Object> notificationMessage;
+    Map<String, Object> notificationTokenId;
 
     private int positionToEdit;
     private FirebaseFirestore db;
@@ -49,18 +60,11 @@ public class EventListFragment extends Fragment {
         View view = inflater.inflate(R.layout.fragment_event_list, container, false);
 
         ListView listView = view.findViewById(R.id.listview_events);
-//        Button btnClose = view.findViewById(R.id.btn_close_event_list);
-
-
         eventDataList = new ArrayList<>();
         eventArrayAdapter = new EventArrayAdapter(getContext(), eventDataList);
         listView.setAdapter(eventArrayAdapter);
-
         fetchData();
 
-//        listView.setOnItemClickListener((adapterView, view1, i, l) -> {
-//            // Handle list item click
-//        });
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
@@ -76,41 +80,39 @@ public class EventListFragment extends Fragment {
             @Override
             public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
                 Event eventToBeDeleted = eventDataList.get(position);
-                // Event(eventId, eventTitle.getText().toString(), "", "",
-                //  new Date(), new GeoPoint(location.getLatitude(), location.getLongitude()), eventId, "", 0)
-                /////////////////////////////////////////
-                message = eventToBeDeleted.getTitle() + " is cancelled.";
-                notificationMessage = new HashMap<>();
-                notificationMessage.put("message", message);
-//                notificationMessage.put("from", )
-                FirebaseUtil.addEventNotification(eventToBeDeleted, notificationMessage,
-                        aVoid -> {
-                            Toast.makeText(getActivity(), "Event cancellation notification has been sent to attnedees", Toast.LENGTH_SHORT).show();
-                        },
-                        e -> {
-                            Toast.makeText(getActivity(), "Failed to send notification", Toast.LENGTH_SHORT).show();
-                        });
-                /////////////////////////////////////////
-                // Call FirebaseUtil.deleteEvent to delete the event
-                FirebaseUtil.deleteEvent(eventToBeDeleted.getId(),
-                        aVoid -> {
-                            // Successfully deleted event from Firestore
-                            eventDataList.remove(position); // Remove event from the local list
-                            eventArrayAdapter.notifyDataSetChanged(); // Notify the adapter to update the list
-                            fetchData();
-                            // Optionally, show a confirmation message
-                            Toast.makeText(getActivity(), "Event " + eventToBeDeleted.getId() + " deleted successfully", Toast.LENGTH_SHORT).show();
 
-                            //Toast.makeText(getActivity(), "Event deleted successfully", Toast.LENGTH_SHORT).show();
-                        },
-                        e -> {
-                            Toast.makeText(getActivity(), "Failed to delete event", Toast.LENGTH_SHORT).show();
-                        });
-                fetchData();
+                // Create an AlertDialog to confirm deletion
+                new AlertDialog.Builder(getActivity())
+                        .setTitle("Delete Event") // Set the title of the dialog
+                        .setMessage("Are you sure you want to delete " + eventToBeDeleted.getTitle() + "?") // Set the message to show to the user
+                        .setPositiveButton("Yes", (dialog, which) -> {
+                            // Notify attendee that event has been cancelled.
+                            String message = eventToBeDeleted.getTitle() + " is cancelled.";
+                            Notification notification = new Notification("notification" + System.currentTimeMillis(), eventToBeDeleted.getId(), message, new Date(), false);
+                            FirebaseUtil.addNotification(notification, aVoid -> {}, e -> {});
+
+                            // Call FirebaseUtil.deleteEvent to delete the event
+                            FirebaseUtil.deleteEvent(eventToBeDeleted.getId(),
+                                    aVoid -> {
+                                        // Successfully deleted event from Firestore
+                                        eventDataList.remove(position); // Remove event from the local list
+                                        eventArrayAdapter.notifyDataSetChanged(); // Notify the adapter to update the list
+                                        fetchData();
+                                        // Optionally, show a confirmation message
+                                        Toast.makeText(getActivity(), "Event " + eventToBeDeleted.getId() + " deleted successfully", Toast.LENGTH_SHORT).show();
+                                    },
+                                    e -> {
+                                        Toast.makeText(getActivity(), "Failed to delete event", Toast.LENGTH_SHORT).show();
+                                    });
+                        })
+                        .setNegativeButton("No", null) // Set the negative button with no additional actions
+                        .setIcon(android.R.drawable.ic_dialog_alert) // Set an icon for the dialog
+                        .show(); // Display the AlertDialog
+
                 return true; // Return true to indicate that the click was handled
             }
-
         });
+
 
         // Close button to go back to the previous screen
 //        btnClose.setOnClickListener(v -> {
@@ -138,7 +140,6 @@ public class EventListFragment extends Fragment {
             @Override
             public void onError(Exception e) {
             }
-
 
         });
     }
